better error handling

+++ eneel/load_runner.py
@@ -1,6 +1,7 @@
 import eneel.utils as utils
 from concurrent.futures import ProcessPoolExecutor as Executor
 import os
+import sys
 import eneel.printer as printer
 import time
 
@@ -12,6 +13,7 @@ def run_project(project_name, connections_path=None):
     # Get configurations
     connections_config = utils.get_connections(connections_path)
     project_config = utils.get_project(project_name)
+
 
     source_name = project_config['source']
     target_name = project_config['target']
@@ -74,39 +76,53 @@ def run_project(project_name, connections_path=None):
                       )
     printer.print_msg('')
 
-
     if num_tables_to_load < workers:
         workers = num_tables_to_load
     start_msg = "Start loading " + str(num_tables_to_load) + " tables with " + str(workers) + " parallel workers"
-    #logger.info("Start loading " + str(num_tables_to_load) + " tables with " + str(workers) + " parallel workers")
     printer.print_output_line(start_msg)
 
     job_start_time = time.time()
 
     # Execute parallel load
+    load_results = []
     with Executor(max_workers=workers) as executor:
-        for _ in executor.map(run_load, load_orders, num_tables_to_loads, project_names, source_conninfos, target_conninfos,
+        for result in executor.map(run_load, load_orders, num_tables_to_loads, project_names, source_conninfos, target_conninfos,
                               projects, temp_paths, schemas, tables):
-            pass
+            load_results.append(result)
+
+    load_errors = 0
+    load_successes = 0
+    for load_result in load_results:
+        if load_result == 'error':
+            load_errors += 1
+        if load_results == 'success':
+            load_successes += 1
 
     # Clean up temp dir
-    utils.delete_path(temp_path)
+    if not project.get('keep_tempfiles', False):
+        utils.delete_path(temp_path)
 
     execution_time = time.time() - job_start_time
 
     status_time = " in {execution_time:0.2f}s".format(
         execution_time=execution_time)
 
-    end_msg = "Finished loading " + str(num_tables_to_load) + " tables in " + status_time
+    end_msg = "Finished loading " + str(load_successes) + " of " + str(num_tables_to_load) + \
+              " tables successfully in " + status_time
     printer.print_output_line("")
     printer.print_output_line(end_msg)
     #logger.info("Finished loading " + str(num_tables_to_load) + " tables ")
 
     printer.print_msg("")
-    printer.print_msg("Completed successfully", "green")
+    if load_errors > 0:
+        printer.print_msg("Completed with errors", "red")
+    else:
+        printer.print_msg("Completed successfully", "green")
 
 
 def run_load(load_order, num_tables_to_load, project_name, source_conninfo, target_conninfo, project, temp_path, schema, table):
+    return_code = 'error'
+
     load_start_time = time.time()
 
     import eneel.logger as logger
@@ -131,6 +147,12 @@ def run_load(load_order, num_tables_to_load, project_name, source_conninfo, targ
     full_source_table = source_schema + '.' + source_table
     target_table = schema.get('table_prefix', "") + table.get('table_name') + schema.get('table_suffix', "")
     full_target_table = target_schema + '.' + target_table
+    index = load_order
+    total = num_tables_to_load
+
+    if not source.check_table_exist(full_source_table):
+        printer.print_load_line(index, total, "ERROR", full_source_table, msg="does not exist in source")
+        return return_code
 
     # Temp table
     target_table_tmp = target_table + '_tmp'
@@ -157,6 +179,7 @@ def run_load(load_order, num_tables_to_load, project_name, source_conninfo, targ
             columns = columns_to_load
     except:
         logger.error("Could not determine columns to load")
+        return return_code
 
     # Load type
     replication_method = table.get('replication_method')
@@ -169,71 +192,82 @@ def run_load(load_order, num_tables_to_load, project_name, source_conninfo, targ
         printer.print_load_line(index, total, status, table_msg)
 
         # Export table
-        file, delimiter = source.export_table(source_schema, source_table, columns, temp_path_load,
+        try:
+            file, delimiter = source.export_table(source_schema, source_table, columns, temp_path_load,
                                               csv_delimiter)
 
-        # Create temp table
-        target.create_table_from_columns(target_schema, target_table_tmp, columns)
-
-        # Import into temp table
-        import_status, import_row_count = target.import_table(target_schema, target_table_tmp, file, delimiter)
-
-        # Switch tables
-        target.switch_tables(target_schema, target_table, target_table_tmp)
-
-    elif replication_method == "INCREMENTAL":
-        index = load_order
-        total = num_tables_to_load
-        status = "START"
-        table_msg = full_source_table + " (" + replication_method + ")"
-        printer.print_load_line(index, total, status, table_msg)
-
-        replication_key = table.get('replication_key')
-
-        if target.check_table_exist(full_target_table) and replication_key:
-            max_replication_key = target.get_max_column_value(full_target_table, replication_key)
-            # Export new rows
-            file, delimiter = source.export_table(source_schema, source_table, columns, temp_path_load, csv_delimiter,
-                                                  replication_key, max_replication_key)
             # Create temp table
             target.create_table_from_columns(target_schema, target_table_tmp, columns)
-            # Import into temp table
-            import_status, import_row_count = target.import_table(target_schema, target_table_tmp, file, delimiter)
-            # Insert into and drop
-            target.insert_from_table_and_drop(target_schema, target_table, target_table_tmp)
-            #logger.info(full_source_table + " updated")
 
-        else:
-            # Full export
-            file, delimiter = source.export_table(source_schema, source_table, columns, temp_path_load, csv_delimiter)
-            # Create temp table
-            target.create_table_from_columns(target_schema, target_table_tmp, columns)
             # Import into temp table
             import_status, import_row_count = target.import_table(target_schema, target_table_tmp, file, delimiter)
+
             # Switch tables
             target.switch_tables(target_schema, target_table, target_table_tmp)
-            logger.info(full_source_table + " imported")
 
-        # Import table
+        except:
+            printer.print_load_line(index, total, "ERROR", full_source_table, msg="load ")
+            return return_code
 
+    elif replication_method == "INCREMENTAL":
+        try:
+            index = load_order
+            total = num_tables_to_load
+            status = "START"
+            table_msg = full_source_table + " (" + replication_method + ")"
+            printer.print_load_line(index, total, status, table_msg)
 
+            replication_key = table.get('replication_key')
+            if not replication_key:
+                printer.print_load_line(index, total, "ERROR", full_source_table, msg="replication key not defined")
+                return return_code
+            if replication_key not in [column[1] for column in columns]:
+                printer.print_load_line(index, total, "ERROR", full_source_table, msg="replication key not found in table")
+                return return_code
+
+            max_replication_key = target.get_max_column_value(full_target_table, replication_key)
+            if not max_replication_key:
+                # Full export
+                file, delimiter = source.export_table(source_schema, source_table, columns, temp_path_load,
+                                                      csv_delimiter)
+                # Create temp table
+                target.create_table_from_columns(target_schema, target_table_tmp, columns)
+                # Import into temp table
+                import_status, import_row_count = target.import_table(target_schema, target_table_tmp, file, delimiter)
+                # Switch tables
+                target.switch_tables(target_schema, target_table, target_table_tmp)
+
+            else:
+                # Export new rows
+                file, delimiter = source.export_table(source_schema, source_table, columns, temp_path_load, csv_delimiter,
+                                                      replication_key, max_replication_key)
+                # Create temp table
+                target.create_table_from_columns(target_schema, target_table_tmp, columns)
+                # Import into temp table
+                import_status, import_row_count = target.import_table(target_schema, target_table_tmp, file, delimiter)
+                # Insert into and drop
+                target.insert_from_table_and_drop(target_schema, target_table, target_table_tmp)
+
+        except:
+            printer.print_load_line(index, total, "ERROR", full_source_table, msg="load failed")
+            return return_code
 
     else:
-        logger.error("replication_method not valid")
+        printer.print_load_line(index, total, "ERROR", full_source_table, msg="replication_method not valid")
+        return return_code
 
     # delete temp folder
-    utils.delete_path(temp_path_load)
+    if not project.get('keep_tempfiles', False):
+        utils.delete_path(temp_path_load)
 
-    end_table_load_msg = "Finished loading: " + full_source_table
-    #printer.print_output_line(end_table_load_msg)
+    # Close connections
+    source.close()
+    target.close()
 
-    #printer.print_fancy_output_line(msg, "OK", 1, 2)
-
-    index = load_order
-    total = num_tables_to_load
     status = import_status
     rows = import_row_count
     execution_time = time.time() - load_start_time
     printer.print_load_line(index, total, status, full_source_table, rows, execution_time)
+    return_code = 'success'
+    return return_code
 
-    #logger.info("Finished loading: " + full_source_table)
